+++ src/prompt.js
@@ -1,10 +1,10 @@
 const chalk = require("chalk");
-const readline = require("readline-sync");
+const { question } = require("readline-sync");
 
 const prompt = () => {
-  let command = readline
-    .question(`${chalk.bold.green("bubble")} ${chalk.blueBright(process.cwd())} ${chalk.red("$")} `)
-    ?.trim();
+  let command = question(
+    `${chalk.bold.green("bubble")} ${chalk.blueBright(process.cwd())} ${chalk.red("$")} `
+  )?.trim();
 
   return command;
 };